+++ rust/operator-binary/src/druid_controller.rs
@@ -443,6 +443,11 @@ fn build_rolegroup_config_map(
 
         match property_name_kind {
             PropertyNameKind::File(file_name) if file_name == RUNTIME_PROPS => {
+                // Add any properties derived from storage manifests, such as segment cache locations.
+                // This has to be done here since there is no other suitable place for it.
+                // Previously such properties were added in the compute_files() function,
+                // but that code path is now incompatible with the design of fragment merging.
+                resources.update_druid_config_file(file_name.as_str(), &mut transformed_config);
                 // NOTE: druid.host can be set manually - if it isn't, the canonical host name of
                 // the local host is used.  This should work with the agent and k8s host networking
                 // but might need to be revisited in the future
@@ -616,6 +621,7 @@ fn build_rolegroup_statefulset(
         &mut cb_druid,
         &mut pb,
     );
+    resources.update_volumes_and_volume_mounts(&mut cb_druid, &mut pb);
 
     let prepare_container_command = tls_settings.build_tls_key_stores_cmd();
 
@@ -787,4 +793,139 @@ fn container_image(version: &str) -> String {
 
 pub fn error_policy(_obj: Arc<DruidCluster>, _error: &Error, _ctx: Arc<Ctx>) -> Action {
     Action::requeue(Duration::from_secs(5))
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use rstest::*;
+    use stackable_druid_crd::PROP_SEGMENT_CACHE_LOCATIONS;
+    use stackable_operator::product_config::{writer, ProductConfigManager};
+
+    #[derive(Snafu, Debug, EnumDiscriminants)]
+    #[strum_discriminants(derive(IntoStaticStr))]
+    #[allow(clippy::enum_variant_names)]
+    pub enum Error {
+        #[snafu(display("controller error"))]
+        Controller { source: super::Error },
+        #[snafu(display("product config error"))]
+        ProductConfig {
+            source: stackable_operator::product_config::error::Error,
+        },
+        #[snafu(display("product config utils error"))]
+        ProductConfigUtils {
+            source: stackable_operator::product_config_utils::ConfigError,
+        },
+        #[snafu(display("operator framework error"))]
+        OperatorFramework {
+            source: stackable_operator::error::Error,
+        },
+        #[snafu(display("resource error"))]
+        Resource {
+            source: stackable_druid_crd::resource::Error,
+        },
+    }
+
+    #[rstest]
+    #[case(
+        "segment_cache.yaml",
+        "default",
+        "[{\"path\":\"/stackable/var/druid/segment-cache\",\"maxSize\":\"1G\",\"freeSpacePercent\":\"5\"}]"
+    )]
+    #[case(
+        "segment_cache.yaml",
+        "secondary",
+        "[{\"path\":\"/stackable/var/druid/segment-cache\",\"maxSize\":\"5G\",\"freeSpacePercent\":\"2\"}]"
+    )]
+    fn segment_cache_location_property(
+        #[case] druid_manifest: &str,
+        #[case] tested_rolegroup_name: &str,
+        #[case] expected_druid_segment_cache_property: &str,
+    ) -> Result<(), Error> {
+        let cluster_cr =
+            std::fs::File::open(format!("test/resources/druid_controller/{druid_manifest}"))
+                .unwrap();
+        let deserializer = serde_yaml::Deserializer::from_reader(&cluster_cr);
+        let druid: DruidCluster =
+            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
+
+        let role_config = transform_all_roles_to_config(&druid, druid.build_role_properties());
+
+        let product_config_manager =
+            ProductConfigManager::from_yaml_file("test/resources/druid_controller/properties.yaml")
+                .context(ProductConfigSnafu)?;
+
+        let validated_role_config = validate_all_roles_and_groups_config(
+            druid.version(),
+            &role_config.context(ProductConfigUtilsSnafu)?,
+            &product_config_manager,
+            false,
+            false,
+        )
+        .context(OperatorFrameworkSnafu)?;
+
+        let druid_tls_settings = DruidTlsSettings {
+            encryption: druid.spec.cluster_config.tls.clone(),
+            authentication: None,
+        };
+
+        let mut druid_segment_cache_property = "invalid".to_string();
+
+        for (role_name, role_config) in validated_role_config.iter() {
+            for (rolegroup_name, rolegroup_config) in role_config.iter() {
+                if rolegroup_name == tested_rolegroup_name
+                    && role_name == &DruidRole::Historical.to_string()
+                {
+                    let rolegroup_ref = RoleGroupRef {
+                        cluster: ObjectRef::from_obj(&druid),
+                        role: role_name.into(),
+                        role_group: rolegroup_name.clone(),
+                    };
+
+                    let resources =
+                        resource::resources(&druid, &DruidRole::Historical, &rolegroup_ref)
+                            .context(ResourceSnafu)?;
+
+                    let rg_configmap = build_rolegroup_config_map(
+                        &druid,
+                        &rolegroup_ref,
+                        rolegroup_config,
+                        "zookeeper-connection-string",
+                        None,
+                        None,
+                        None,
+                        &resources,
+                        &druid_tls_settings,
+                    )
+                    .context(ControllerSnafu)?;
+
+                    druid_segment_cache_property = rg_configmap
+                        .data
+                        .unwrap()
+                        .get(&RUNTIME_PROPS.to_string())
+                        .unwrap()
+                        .to_string();
+
+                    break;
+                }
+            }
+        }
+        let escaped_segment_cache_property = writer::to_java_properties_string(
+            vec![(
+                &PROP_SEGMENT_CACHE_LOCATIONS.to_string(),
+                &Some(expected_druid_segment_cache_property.to_string()),
+            )]
+            .into_iter(),
+        )
+        .unwrap();
+
+        //assert!(druid_segment_cache_property.contains(expected_druid_segment_cache_property), "role group {}", tested_rolegroup_name);
+        assert!(
+            druid_segment_cache_property.contains(&escaped_segment_cache_property),
+            "role group {}",
+            tested_rolegroup_name
+        );
+
+        Ok(())
+    }
 }